Update on Fab Permission

+++ app/Http/Controllers/BAAController.php
@@ -11,6 +11,7 @@ use App\Models\SPK;
 use App\Models\User;
 use App\Service\BAAService;
 use App\Service\SPKService;
+use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Http\JsonResponse;
 use Illuminate\Http\Request;
 use Illuminate\Http\Response;
@@ -32,49 +33,81 @@ class BAAController extends Controller
         $this->user = new User();
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function index(): View
     {
+        $this->authorize('viewAny', BAA::class);
         return view('pages.transaction.baa.index');
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function data(): JsonResponse
     {
+        $this->authorize('view', BAA::class);
         return response()->json($this->baaService->data());
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function search(Request $request): JsonResponse
     {
+        $this->authorize('view', BAA::class);
         return response()->json($this->baaService->search($request));
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function getFabData(Request $request): JsonResponse
     {
+        $this->authorize('view', BAA::class);
         return response()->json($this->fab->getData($request));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function selectedFabData(BAA $baa): JsonResponse
     {
+        $this->authorize('update', $baa);
         return response()->json($this->fab->getSelectedData($baa->fab_id));
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function create(): View
     {
+        $this->authorize('create', BAA::class);
         return view('pages.transaction.baa.create');
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function store(BaaRequest $request): JsonResponse
     {
+        $this->authorize('create', BAA::class);
         $this->baaService->store($request);
         return response()->json(['message' => 'data berhasil disimpan']);
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function detail(BAA $baa): View
     {
+        $this->authorize('viewDetail', $baa);
         $getFabServiceCategory = FabServiceCategory::with('service')->where('fab_id', $baa->fab_id)->get();
 
 
@@ -89,19 +122,31 @@ class BAAController extends Controller
         return view('pages.transaction.baa.detail', compact('baa', 'serviceCategory'));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function edit(BAA $baa): View
     {
+        $this->authorize('update', $baa);
         return view('pages.transaction.baa.edit', compact('baa'));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function update(BaaRequest $request, Baa $baa): JsonResponse
     {
+        $this->authorize('update', $baa);
         $this->baaService->update($request, $baa);
         return response()->json(['message' => 'data berhasil disimpan']);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function confirm(BAA $baa): JsonResponse
     {
+        $this->authorize('confirm', $baa);
         $baa->update([
             'status' => 1
         ]);
@@ -110,14 +155,22 @@ class BAAController extends Controller
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function destroy(BAA $baa): JsonResponse
     {
+        $this->authorize('delete', $baa);
         $baa->delete();
         return response()->json(['message' => 'data berhasil dihapus']);
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function exportPDF(BAA $baa): Response
     {
+        $this->authorize('print', $baa);
         $view = view('pages.transaction.baa.export-pdf', compact('baa'))->render();
         $pdf = Browsershot::html($view)
             ->setChromePath('/usr/bin/chromium')
@@ -136,13 +189,21 @@ class BAAController extends Controller
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function getUserData(Request $request): JsonResponse
     {
+        $this->authorize('create', BAA::class);
         return response()->json($this->user->getUser($request));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function saveSpk(SPKRequest $request, BAA $baa): JsonResponse
     {
+        $this->authorize('createOrUpdateSPK', $baa);
         SPK::updateOrCreate([
             'baa_id' => $baa->id,
         ], [
@@ -159,25 +220,40 @@ class BAAController extends Controller
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function getSpk(BAA $baa): JsonResponse
     {
+        $this->authorize('viewDetail', $baa);
         return response()->json(SPK::where('baa_id', $baa->id)->first());
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function getSelectedFrom(User $user): JsonResponse
     {
+        $this->authorize('update', BAA::class);
         return response()->json($this->user->getSelectedData($user->id));
     }
 
 
+    /**
+     * @throws AuthorizationException
+     */
     public function getSelectedTo(User $user): JsonResponse
     {
+        $this->authorize('update', BAA::class);
         return response()->json($this->user->getSelectedData($user->id));
     }
 
+    /**
+     * @throws AuthorizationException
+     */
     public function exportSpkToPDF(BAA $baa): Response
     {
-
+        $this->authorize('printSPK', $baa);
         $spk = SPK::where('baa_id', $baa->id)->first();
         $view = view('pages.transaction.baa.spk.export-pdf', compact('baa', 'spk'))->render();
         $pdf = Browsershot::html($view)
